refactor(core): move auth from ee/ to auth/ folder
- Move all auth features from packages/core/src/ee/ to packages/core/src/auth/
- Change import path from @mastra/core/ee to @mastra/core/auth
- Rename isEELicenseValid() to isLicenseValid() (with backward compat alias)
- Remove unused withEE wrapper (dead code)
- Update all internal and external imports across auth packages,
  server adapters, and server handlers

This decouples folder organization from licensing model.
