+++ auth/better-auth/src/index.ts
@@ -1,4 +1,4 @@
-import type { IUserProvider, ICredentialsProvider, EEUser, CredentialsResult } from '@mastra/core/ee';
+import type { IUserProvider, ICredentialsProvider, EEUser, CredentialsResult } from '@mastra/core/auth';
 import type { MastraAuthProviderOptions } from '@mastra/core/server';
 import { MastraAuthProvider } from '@mastra/core/server';
