+++ packages/core/src/auth/defaults/index.ts
