+++ packages/core/src/auth/defaults/rbac/static.ts
