+++ packages/core/src/auth/defaults/session/cookie.ts
