+++ packages/core/src/auth/defaults/session/index.ts
