+++ packages/core/src/auth/interfaces/index.ts
