+++ packages/core/src/auth/interfaces/permissions.generated.ts
