+++ packages/core/src/auth/interfaces/session.ts
