+++ packages/core/src/auth/interfaces/user.ts
