+++ server-adapters/hono/src/index.ts
@@ -1,5 +1,5 @@
 import type { ToolsInput } from '@mastra/core/agent';
-import { hasPermission } from '@mastra/core/ee';
+import { hasPermission } from '@mastra/core/auth';
 import type { Mastra } from '@mastra/core/mastra';
 import type { RequestContext } from '@mastra/core/request-context';
 import type { InMemoryTaskStore } from '@mastra/server/a2a/store';
